// types/icecream-category.ts
export type IceCreamCategory = {
    id: number
    name: string
    sugarsMin: number
    sugarsMax: number
    fatsMin: number
    fatsMax: number
    msnfMin: number
    msnfMax: number
    otherSolidsMin: number
    otherSolidsMax: number
    totalSolidsMin: number
    totalSolidsMax: number
    podMin: number
    podMax: number
    pacMin: number
    pacMax: number
    fruitMin: number
    fruitMax: number
    alcoholMin: number
    alcoholMax: number
    overrunMin: number
    overrunMax: number
    groundFoodsMin: number
    groundFoodsMax: number
}